+++ spec/spec_helper.rb
@@ -12,15 +12,10 @@ unless defined?(SPEC_ROOT)
   SPEC_ROOT = File.expand_path("../", __FILE__)
 end
 
-if ENV['DNSIMPLE_TEST_CONFIG']
-  DNSimple::Client.load_credentials(ENV['DNSIMPLE_TEST_CONFIG'])
-  CONFIG = { 'username' => DNSimple::Client.username, 'password' => DNSimple::Client.password, 'base_uri' => DNSimple::Client.base_uri, 'host' => URI.parse(DNSimple::Client.base_uri).host }
-else
-  CONFIG = { 'username' => 'username', 'password' => 'password', 'base_uri' => 'https://api.sandbox.dnsimple.com/', 'host' => 'api.sandbox.dnsimple.com' }
-  DNSimple::Client.base_uri = CONFIG['base_uri']
-  DNSimple::Client.username = CONFIG['username']
-  DNSimple::Client.password = CONFIG['password']
-end
+CONFIG = { 'username' => 'username', 'password' => 'password', 'base_uri' => 'https://api.sandbox.dnsimple.com/', 'host' => 'api.sandbox.dnsimple.com' }
+DNSimple::Client.base_uri = CONFIG['base_uri']
+DNSimple::Client.username = CONFIG['username']
+DNSimple::Client.password = CONFIG['password']
 
 
 RSpec.configure do |c|